% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AlphaPalettes}
\alias{AlphaPalettes}
\title{Alpha palettes by Tinkerbelly}
\format{
\subsection{\code{AlphaPalettes}}{

A list of named vectors
\describe{
\item{name}{Palette name}
\item{values}{Vector of alpha values}
}
}
}
\source{
Own preferences
}
\usage{
AlphaPalettes
}
\description{
A list of named vectors containing alpha values
}
\author{
Bastian Reiter
}
\keyword{datasets}
