% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCPSiteCredentials}
\alias{CCPSiteCredentials}
\title{Template of a data frame containing login credentials being passed to \code{\link{ConnectToCCP}}.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 3 columns.
}
\usage{
CCPSiteCredentials
}
\description{
Data frame with three columns
\itemize{\item{SiteName}
\item{URL}
\item{Token}}
}
\author{
Bastian Reiter
}
\keyword{datasets}
