% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCPSiteSpecifications}
\alias{CCPSiteSpecifications}
\title{Template of a data frame containing site-specific login credentials and project names being passed to \code{\link{ConnectToCCP}}.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.
}
\usage{
CCPSiteSpecifications
}
\description{
Data frame with four columns
\itemize{\item{SiteName}
\item{URL}
\item{ProjectName}
\item{Token}}
}
\author{
Bastian Reiter
}
\keyword{datasets}
