% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Colors}
\alias{Colors}
\title{Preferred colors by Tinkerbelly}
\format{
\subsection{\code{Colors}}{

A list of named strings (hexadecimal RGB codes)
\describe{
\item{name}{Color name}
\item{code}{Color code}
}
}
}
\source{
Own preferences after researching a lot of color palette stuff
}
\usage{
Colors
}
\description{
A list of named hexadecimal RGB codes
}
\author{
Bastian Reiter
}
\keyword{datasets}
