% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToVirtualCCP.R
\name{ConnectToVirtualCCP}
\alias{ConnectToVirtualCCP}
\title{ConnectToVirtualCCP}
\usage{
ConnectToVirtualCCP(
  CCPData,
  CCPData.Mode = "Auto",
  NumberOfServers = 1,
  NumberOfPatientsPerServer = NULL,
  AddedDsPackages = NULL,
  Resources = NULL,
  WorkingDirectory = file.path(tempdir(), ".dslite")
)
}
\arguments{
\item{CCPData}{\code{Named list} of \code{data.frames} - CCP test data}

\item{CCPData.Mode}{\code{string} - One of:
\itemize{ \item 'Single' - Data in 'CCPData' contains data from a single site (or test data) that can be distributed randomly between virtual servers
\item 'Multi' - Data in 'CCPData' contains a list with multi-site data
\item 'Auto' - Functions tries to determine from structure of 'CCPData' which mode should be used (Default) }}

\item{NumberOfServers}{\code{integer} - The number of virtual servers to install}

\item{NumberOfPatientsPerServer}{\code{integer} - Optional value to restrict size of data set for faster testing - Default: NULL}

\item{AddedDsPackages}{\code{character vector} - Server-side DataSHIELD packages to be added to default (dsBase, dsCCPhos) - Default: NULL}

\item{Resources}{\code{Named list} of \code{resourcer::Resource} objects - Default: NULL}

\item{WorkingDirectory}{\code{string} - Optional custom working directory for virtual servers - Default: Hidden folder in R session's temporary directory (see \code{?DSLite::newDSLiteServer()})}
}
\value{
A \code{list} of \code{DSConnection}-objects
}
\description{
Sets up a virtual DataSHIELD infrastructure that enables trying out real dsCCPhos functionality on test data.
}
\author{
Bastian Reiter
}
