% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisplayCrossTable.R
\name{DisplayCrossTable}
\alias{DisplayCrossTable}
\title{DisplayCrossTable}
\usage{
DisplayCrossTable(
  CrossTabData,
  FeatureX,
  FeatureY,
  IncludeMissingTimePoints = FALSE
)
}
\arguments{
\item{CrossTabData}{\code{data.frame} - Containing data output by \code{dsFredaClient::ds.GetCrossTab()}}

\item{FeatureX}{\code{symbol} or \code{string} - Feature in 'TimeSeriesData' that contains time points (e.g. years)}

\item{FeatureY}{\code{symbol} or \code{string} - Feature in 'TimeSeriesData' that contains values}

\item{IncludeMissingTimePoints}{\code{logical} - Indicates whether discrete time points (e.g. years) with no data should be displayed in the time series - Default: \code{FALSE}}
}
\value{
A \code{data.frame} displaying time series data in horizontal form
}
\description{
Transform CrossTab data output by \code{dsFredaClient::ds.GetCrossTab()} into display-friendly format.
}
\author{
Bastian Reiter
}
