% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.AugmentData.R
\name{ds.AugmentData}
\alias{ds.AugmentData}
\title{ds.AugmentData}
\usage{
ds.AugmentData(
  Name_CurationOutput = "CurationOutput",
  Name_Output = "AugmentationOutput",
  DataSources = NULL
)
}
\arguments{
\item{Name_Output}{String | Name of assigned output object on server | Default: 'CurationOutput'}

\item{DataSources}{List of DSConnection objects}

\item{Name_RawData}{String | Name of raw data object (list) on server | Default: 'RawData'}
}
\value{
A list of variables containing messages about object assignment for monitoring purposes.
}
\description{
What it does
}
\details{
Linked to server-side ASSIGN method AugmentDataDS()
}
