% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.CheckDataSet.R
\name{ds.CheckDataSet}
\alias{ds.CheckDataSet}
\title{ds.CheckDataSet}
\usage{
ds.CheckDataSet(
  DataSources = NULL,
  DataSetName,
  RequiredTableNames = NULL,
  RequiredFeatureNames = NULL
)
}
\arguments{
\item{DataSources}{\code{list} of \code{DSConnection} objects}

\item{DataSetName}{\code{string} - Name of Data Set object (list) on server, usually "RawDataSet", "CuratedDataSet" or "AugmentedDataSet"}

\item{RequiredTableNames}{\code{character vector} - Names of tables that are expected/required to be in the data set - Default: Names of elements in list evaluated from \code{DataSetName.S}}

\item{RequiredFeatureNames}{\code{list} of \code{character vectors} - Features that are expected/required in each table of the data set - Default: Names of features in respective table}
}
\value{
A \code{list} containing compiled meta data about RDS table:
\itemize{ \item TableStatus
\item TableRowCounts
\item FeatureStatus
\item FeatureTypes
\item NonMissingValueRates }
}
\description{
Check out a data set (list of data.frames) on servers and return a coherent summary across servers.
}
\details{
Linked to server-side AGGREGATE method CheckDataSetDS()
}
\author{
Bastian Reiter
}
