% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.CheckRDSTables.R
\name{ds.CheckRDSTables}
\alias{ds.CheckRDSTables}
\title{ds.CheckRDSTables}
\usage{
ds.CheckRDSTables(DataSources = NULL, RawDataSetName = "RawDataSet")
}
\arguments{
\item{DataSources}{List of DSConnection objects}
}
\value{
A list
}
\description{
Check Raw Data Set (RDS) tables on servers on existence and completeness and return a coherent summary across servers.
}
\details{
Linked to server-side AGGREGATE method CheckRDSTablesDS()
}
\author{
Bastian Reiter
}
