% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.CheckSiteRequirements.R
\name{ds.CheckSiteRequirements}
\alias{ds.CheckSiteRequirements}
\title{ds.CheckSiteRequirements}
\usage{
ds.CheckSiteRequirements(DataSources)
}
\arguments{
\item{Name_CurationOutput}{data.frame or tibble | Data on CCP site credentials}
}
\description{
Check if technical requirements are met by every participating CCP site.
}
