% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.ConnectToCCP.R
\name{ds.ConnectToCCP}
\alias{ds.ConnectToCCP}
\title{ds.ConnectToCCP}
\usage{
ds.ConnectToCCP(CCPSiteCredentials)
}
\arguments{
\item{CCPSiteCredentials}{data.frame or tibble | Data on CCP site credentials}
}
\value{
A list of DSConnection-objects
}
\description{
Takes credentials of CCP sites and returns a list of DSConnection-objects. Has to be executed from within a CCP bridgehead R server session.
}
