% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.CreateCheckpoints.R
\name{ds.CreateCheckpoints}
\alias{ds.CreateCheckpoints}
\title{ds.CreateCheckpoints}
\usage{
ds.CreateCheckpoints(DataSources = NULL)
}
\arguments{
\item{DataSources}{List of DSConnection objects}
}
\value{
A list of messages about object assignment for monitoring purposes
}
\description{
Triggers assignment of named vector on servers containing status data on processing checkpoints
}
\details{
Linked to server-side ASSIGN method CreateCheckpointsDS()
}
\author{
Bastian Reiter
}
