% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.CurateData.R
\name{ds.CurateData}
\alias{ds.CurateData}
\title{ds.CurateData}
\usage{
ds.CurateData(
  Name_RawData = "RawData",
  Name_Output = "CurationOutput",
  DataSources = NULL
)
}
\arguments{
\item{Name_RawData}{String | Name of raw data object (list) on server | Default: 'RawData'}

\item{Name_Output}{String | Name of assigned output object on server | Default: 'CurationOutput'}

\item{DataSources}{}
}
\description{
What it does
}
\details{
Linked to server-side ASSIGN method CurateDataDS()
}
