% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.DrawSample.R
\name{ds.DrawSample}
\alias{ds.DrawSample}
\title{ds.DrawSample}
\usage{
ds.DrawSample(
  RawDataSetName = "RawDataSet",
  SampleSize = "10",
  SampleName = "RawDataSet",
  DataSources = NULL
)
}
\arguments{
\item{RawDataSetName}{\code{string} | Name of a list object on server}

\item{SampleSize}{\code{string} | Number of patients per site}

\item{SampleName}{\code{string} | Option to assign subset of Raw Data Set a different object name on servers}

\item{DataSources}{List of DSConnection objects}
}
\value{
A list of messages about object assignment for monitoring purposes
}
\description{
Draw a random sample from Raw Data Set
}
\details{
Linked to server-side ASSIGN method DrawSampleDS()
}
\author{
Bastian Reiter
}
