% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.ExtractFromList.R
\name{ds.ExtractFromList}
\alias{ds.ExtractFromList}
\title{ds.ExtractFromList}
\usage{
ds.ExtractFromList(
  ListName,
  ObjectName,
  AssignedObjectName = NULL,
  DataSources = NULL
)
}
\arguments{
\item{ListName}{String | Name of a list object on server}

\item{ObjectName}{String | Name of object inside list}

\item{DataSources}{List of DSConnection objects}

\item{NewObjectName}{String | Optionally assigned name of object after extraction to server session | Default: NULL}
}
\value{
A list of messages about object assignment for monitoring purposes
}
\description{
Auxiliary function to trigger assignment of object extracted from a list on server
}
\details{
Linked to server-side ASSIGN method ExtractFromListDS()
}
\author{
Bastian Reiter
}
