% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetCrossTable.R
\name{ds.GetCrossTab}
\alias{ds.GetCrossTab}
\title{ds.GetCrossTab}
\usage{
ds.GetCrossTab(TableName, FeatureNames, DSConnections = NULL)
}
\arguments{
\item{TableName}{\code{string} - Name of \code{data.frame} on server}

\item{FeatureNames}{\code{character} - Vector of feature names}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list}
}
\description{
Get cross tab
}
\details{
Linked to server-side \code{AGGREGATE} function \code{GetCrossTabDS()}
}
\author{
Bastian Reiter
}
