% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetFeatureInfo.R
\name{ds.GetFeatureInfo}
\alias{ds.GetFeatureInfo}
\title{ds.GetFeatureInfo}
\usage{
ds.GetFeatureInfo(DataSources = NULL, TableName, FeatureName)
}
\arguments{
\item{DataSources}{\code{list} of DSConnection objects}

\item{TableName}{\code{string} | Name of the table containing the feature of concern}

\item{FeatureName}{\code{string} | Name of feature}
}
\value{
A \code{tibble} containing separate and cumulated meta data about feature type and sample size.
}
\description{
Obtain data about feature type and sample size.
}
\details{
Linked to server-side \code{AGGREGATE} function \code{dsCCPhos::GetFeatureInfoDS()}.
}
\author{
Bastian Reiter
}
