% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetFrequencyTable.R
\name{ds.GetFrequencyTable}
\alias{ds.GetFrequencyTable}
\title{ds.GetFrequencyTable}
\usage{
ds.GetFrequencyTable(
  DataSources = NULL,
  TableName,
  FeatureName,
  GroupingFeatureName = NULL,
  MaxNumberCategories = NULL
)
}
\arguments{
\item{DataSources}{List of DSConnection objects}

\item{TableName}{String | Name of the table containing the feature of concern}

\item{FeatureName}{String | Name of feature}

\item{GroupingFeatureName}{String | Name of optional grouping feature from the same table}

\item{MaxNumberCategories}{\code{integer} | Maximum number of categories analyzed individually before frequencies are cumulated in 'Other' category.}
}
\value{
A \code{list} containing:
\itemize{\item AbsoluteFrequencies (\code{tibble}: Absolute value frequencies)
\item RelativeFrequencies (\code{tibble}: Relative value frequencies)}
}
\description{
Get table of absolute and relative value frequencies for a nominal / ordinal feature.
}
\details{
Linked to server-side \code{AGGREGATE} function \code{dsCCPhos::GetFrequencyTableDS()}.
}
\author{
Bastian Reiter
}
