% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetObjectInfo.R
\name{ds.GetObjectInfo}
\alias{ds.GetObjectInfo}
\title{ds.GetObjectInfo}
\usage{
ds.GetObjectInfo(ObjectName, DataSources)
}
\arguments{
\item{ObjectName}{String | Name of object on server}

\item{DataSources}{List of DSConnection objects}
}
\description{
What it does
}
\details{
Linked to server-side AGGREGATE method GetObjectInfoDS()
}
