% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetObjectStatus.R
\name{ds.GetObjectStatus}
\alias{ds.GetObjectStatus}
\title{ds.GetObjectStatus}
\usage{
ds.GetObjectStatus(ObjectName, DataSources)
}
\arguments{
\item{ObjectName}{String | Name of object on server}

\item{DataSources}{List of DSConnection objects}
}
\description{
Checks if an object exists on every server in a valid form and returns appropriate messages.
}
\details{
Linked to server-side AGGREGATE method GetObjectStatusDS()
}
\author{
Bastian Reiter
}
