% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetRDSTableCheck.R
\name{ds.GetRDSTableCheck}
\alias{ds.GetRDSTableCheck}
\title{ds.GetRDSTableCheck}
\usage{
ds.GetRDSTableCheck(DataSources = NULL, RawDataSetName = "RawDataSet")
}
\arguments{
\item{DataSources}{List of DSConnection objects}

\item{RawDataSetName}{String | Name of Raw Data Set object (list) on server | Default: 'RawDataSet'}
}
\value{
A list of reporting objects
}
\description{
Triggers a function that checks raw data set (RDS) tables on servers.
}
\details{
Linked to server-side AGGREGATE method GetRDSTableCheckDS()
}
\author{
Bastian Reiter
}
