% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetReportingObject.R
\name{ds.GetReportingObject}
\alias{ds.GetReportingObject}
\title{ds.GetReportingObject}
\usage{
ds.GetReportingObject(ObjectName, DataSources = NULL)
}
\arguments{
\item{ObjectName}{String | Name of reporting object on server. Must be on a list of permitted object names.}

\item{DataSources}{List of DSConnection objects}
}
\value{
A list of reporting objects
}
\description{
Receives a reporting object from servers. Its name must be on a list of permitted object names to ensure data privacy.
}
\details{
Linked to server-side AGGREGATE method GetReportingObjectDS()
}
\author{
Bastian Reiter
}
