% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetSampleStatistics.R
\name{ds.GetSampleStatistics}
\alias{ds.GetSampleStatistics}
\title{ds.GetSampleStatistics}
\usage{
ds.GetSampleStatistics(
  TableName,
  MetricFeatureName,
  GroupingFeatureName = NULL,
  DataSources = NULL
)
}
\arguments{
\item{TableName}{String | Name of the table containing the metric feature of concern}

\item{MetricFeatureName}{String | Name of metric feature}

\item{GroupingFeatureName}{String | Name of optional grouping feature in the same table}

\item{DataSources}{List of DSConnection objects}
}
\description{
Making use of dsBaseClient::ds.meanSdGp() and dsBaseClient::ds.quantileMean() to provide common parametric and nonparametric statistics about a metric feature.
}
\details{
Linked to server-side AGGREGATE dsBase-functions
}
\author{
Bastian Reiter
}
