% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetSurvModel.R
\name{ds.GetSurvModel}
\alias{ds.GetSurvModel}
\title{ds.GetSurvModel}
\usage{
ds.GetSurvModel(
  DataSources = NULL,
  TableName,
  TimeFeature,
  EventFeature,
  CovariateA = NULL,
  CovariateB = NULL,
  CovariateC = NULL,
  MinFollowUpTime = 1
)
}
\arguments{
\item{DataSources}{List of DSConnection objects}

\item{TableName}{\code{string} | Name of the table containing the features of concern}

\item{TimeFeature}{\code{string} | Name of time feature}

\item{EventFeature}{\code{string} | Name of event feature}

\item{CovariateA}{\code{string} | Name of optional CovariateA}

\item{CovariateB}{\code{string} | Name of optional CovariateB}

\item{CovariateC}{\code{string} | Name of optional CovariateC}

\item{MinFollowUpTime}{\code{integer} | Optional minimum of observed follow up time}
}
\value{
A list of survival models
}
\description{
Get Survival
}
\details{
Linked to server-side \code{AGGREGATE} function \code{dsCCPhos::GetSurvModelDS()}.
}
\author{
Bastian Reiter
}
