% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetValidationReport_AugmentedData.R
\name{ds.GetValidationReport_AugmentedData}
\alias{ds.GetValidationReport_AugmentedData}
\title{ds.GetValidationReport_AugmentedData}
\usage{
ds.GetValidationReport_AugmentedData(
  Name_AugmentedDataSet = "AugmentedDataSet",
  DataSources = NULL
)
}
\arguments{
\item{Name_AugmentedDataSet}{String | Name of Augmented Data Set object (list) on server | Default: 'AugmentedDataSet'}

\item{DataSources}{List of DSConnection objects}
}
\value{
A list of tibbles containing output of validation
}
\description{
Triggers validation of Augmented Data Set on server and requests a report.
}
\details{
Linked to server-side AGGREGATE method GetValidationReportDS_AugmentedData()
}
