% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetValidationReport_RawData.R
\name{ds.GetValidationReport_RawData}
\alias{ds.GetValidationReport_RawData}
\title{ds.GetValidationReport_RawData}
\usage{
ds.GetValidationReport_RawData(
  Name_RawDataSet = "RawDataSet",
  DataSources = NULL
)
}
\arguments{
\item{Name_RawDataSet}{String | Name of Raw Data Set object (list) on server | Default: 'RawDataSet'}

\item{DataSources}{List of DSConnection objects}
}
\value{
A list of tibbles containing output of validation
}
\description{
Triggers validation of Raw Data Set on server and requests a report.
}
\details{
Linked to server-side AGGREGATE method GetValidationReportDS_RawData()
}
