% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.UnpackAugmentedDataSet.R
\name{ds.UnpackAugmentedDataSet}
\alias{ds.UnpackAugmentedDataSet}
\title{ds.UnpackAugmentedDataSet}
\usage{
ds.UnpackAugmentedDataSet(
  Name_AugmentationOutput = "AugmentationOutput",
  DataSources = NULL
)
}
\arguments{
\item{Name_AugmentationOutput}{String | Name of Augmentation Output object (list) on server | Default: 'AugmentationOutput'}

\item{DataSources}{List of DSConnection objects}
}
\value{
A list of messages about object assignment for monitoring purposes
}
\description{
Make tables within Augmented Data Set (which lives as a list within Augmentation Output) directly addressable in R server sessions
}
\details{
Linked to server-side ASSIGN method UnpackAugmentedDataSetDS()
}
\author{
Bastian Reiter
}
