% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.UnpackCuratedDataSet.R
\name{ds.UnpackCuratedDataSet}
\alias{ds.UnpackCuratedDataSet}
\title{ds.UnpackCuratedDataSet}
\usage{
ds.UnpackCuratedDataSet(
  Name_CurationOutput = "CurationOutput",
  DataSources = NULL
)
}
\arguments{
\item{Name_CurationOutput}{String | Name of curation output object (list) on server | Default: 'CurationOutput'}

\item{DataSources}{List of DSConnection objects}
}
\value{
A list of messages about object assignment for monitoring purposes
}
\description{
Make tables within Curated Data Set (which lives as a list within Curation Output) directly addressable in R server sessions
}
\details{
Linked to server-side ASSIGN method UnpackCuratedDataSetDS()
}
\author{
Bastian Reiter
}
