% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.UpdateCCPhos.R
\name{ds.UpdateCCPhos}
\alias{ds.UpdateCCPhos}
\title{ds.UpdateCCPhos}
\usage{
ds.UpdateCCPhos(DataSources = NULL)
}
\arguments{
\item{DataSources}{List of DSConnection objects}
}
\value{
A message ...
}
\description{
Triggers re-installing of package 'dsCCPhos' on server
}
\details{
Linked to server-side ASSIGN method UpdateCCPhosDS()
}
