% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploreFeature.R
\name{ExploreFeature}
\alias{ExploreFeature}
\title{ExploreFeature}
\usage{
ExploreFeature(DataSources = NULL, TableName, FeatureName, ...)
}
\arguments{
\item{DataSources}{List of DSConnection objects}

\item{TableName}{\code{string} | Name of the table containing the feature of concern}

\item{FeatureName}{\code{string} | Name of feature}
}
\description{
Get characterizing statistics about a feature of arbitrary data type.
}
\author{
Bastian Reiter
}
