% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetServerCheckpoints.R
\name{GetServerCheckpoints}
\alias{GetServerCheckpoints}
\title{GetServerCheckpoints}
\usage{
GetServerCheckpoints(CCPSiteSpecifications, DataSources = NULL)
}
\arguments{
\item{DataSources}{List of DSConnection objects}
}
\value{
A list
}
\description{
Collect server status info with respect to defined checkpoints in preparation and processing.
}
\author{
Bastian Reiter
}
