% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetServerWorkspaceInfo.R
\name{GetServerWorkspaceInfo}
\alias{GetServerWorkspaceInfo}
\title{GetServerWorkspaceInfo}
\usage{
GetServerWorkspaceInfo(DataSources = NULL)
}
\arguments{
\item{DataSources}{List of DSConnection objects}
}
\value{
A data frame
}
\description{
Check which objects live in server-side R sessions
}
\author{
Bastian Reiter
}
