% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTheme_CCP.R
\name{ggTheme_CCP}
\alias{ggTheme_CCP}
\title{ggTheme_CCP}
\usage{
ggTheme_CCP(
  ...,
  BaseSize = 11,
  LegendPosition = "right",
  SizeFactorPlotTitle = 1.67,
  SizeFactorPlotSubtitle = 1.33,
  SizeFactorPlotCaption = 1,
  SizeFactorAxisLabels = 1.33,
  SizeFactorTickLabels_x = 1,
  SizeFactorTickLabels_y = 1,
  SizeFactorLegendLabels = 1,
  SizeFactorFacetLabels = 1.33
)
}
\arguments{
\item{BaseSize}{Font size that relatively determines all other sizes}

\item{LegendPosition}{Legend position ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{SizeFactorPlotTitle}{Determines font size of plot title}

\item{SizeFactorFacetLabels}{}
}
\description{
Custom ggplot2 theme, based on ggplot2::theme_classic()
}
\author{
Bastian Reiter
}
