% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadRawDataSet.R
\name{LoadRawDataSet}
\alias{LoadRawDataSet}
\title{LoadRawDataSet}
\usage{
LoadRawDataSet(DataSources, ProjectName = "Virtual")
}
\arguments{
\item{DataSources}{List of DSConnection objects}

\item{ProjectName}{Name of project as stated in Opal. Use Default "Virtual" for virtual CCP connection.}
}
\value{
A list of messages
}
\description{
Load raw data set from Opal data base into R session on servers.
}
\author{
Bastian Reiter
}
