% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeBoxPlot.R
\name{MakeBoxPlot}
\alias{MakeBoxPlot}
\title{MakeBoxPlot}
\usage{
MakeBoxPlot(
  SampleStatistics,
  XFeature,
  XFeatureSelection = NULL,
  YFeature,
  OutlierQuantile = 1,
  OutlierAcrossAll = TRUE,
  LogTransform = FALSE,
  AxisLimits_y = c(NA_integer_, NA_integer_),
  AxisTitle_x = "",
  AxisTitle_y = "",
  TickLabelWidth_x = 10,
  Decimals = 0,
  ggTheme = function(...) dsCCPhosClient::ggTheme_CCP(...),
  ThemeArguments = list(),
  FillPalette = NULL,
  ...
)
}
\arguments{
\item{XFeature}{Contains X Values}

\item{XFeatureSelection}{Optional: Character vector that determines selection and order of categories}

\item{YFeature}{Contains Y Values}

\item{OutlierQuantile}{If set on "1", no outliers are cut}

\item{OutlierAcrossAll}{If TRUE, outliers will be calculated across all subgroups}

\item{LogTransform}{Optional: Whether Y data should be transformed using natural logarithm}

\item{AxisLimits_y}{Auto y axis limits as default}

\item{Decimals}{Number of decimals of y axis number format}

\item{ggTheme}{Pass custom theme}

\item{ThemeArguments}{Pass custom theme arguments}

\item{...}{}

\item{DataFrame}{The data frame containing plot data}

\item{ShowViolinPlot}{Optional: Control visibility of violin plot}
}
\value{
a ggplot2 object
}
\description{
MakeBoxPlot
}
