% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifyConnections.R
\name{ModifyConnections}
\alias{ModifyConnections}
\title{ModifyConnections}
\usage{
ModifyConnections(DataSources, ServersToBeRemoved = NULL)
}
\arguments{
\item{DataSources}{List of DSConnection objects}

\item{ServersToBeRemoved}{\code{character vector} of server names to be removed from connections}
}
\value{
A (modified) list of DSConnection-objects
}
\description{
Modifies current list of DSConnection-objects created by DataSHIELD functionality or ConnectToCCP() respectively.
}
\author{
Bastian Reiter
}
