% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{PrintMessages}
\alias{PrintMessages}
\title{PrintMessages}
\usage{
PrintMessages(Messages)
}
\arguments{
\item{Text}{\code{list} List of Sublists of named vectors}

\item{IsClassInfo}{\code{logical}}

\item{IsClassSuccess}{\code{logical}}

\item{IsClassWarning}{\code{logical}}

\item{IsClassFailure}{\code{logical}}
}
\description{
Take list of messages, add symbols and print them on console
}
\keyword{internal}
