% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCP.PreprocessingQuickRun.R
\name{CCP.PreprocessingQuickRun}
\alias{CCP.PreprocessingQuickRun}
\title{CCP.PreprocessingQuickRun}
\usage{
CCP.PreprocessingQuickRun(
  ServerSpecifications = NULL,
  GetReports = TRUE,
  RunAssignmentChecks = FALSE,
  RunSeparately = FALSE,
  DSConnections = NULL,
  DS.async = dsFredaClient::Set.DSSettings$DS.async
)
}
\arguments{
\item{ServerSpecifications}{\code{data.frame} - Same \code{data.frame} used for login. Used here only for acquisition of server-specific project names (in case they are differing) - Default: \code{NULL} for virtual project}

\item{GetReports}{\code{logical} - Indicating whether reporting objects should be obtained from the servers. Can be set to \code{FALSE} to save time. - Default: \code{TRUE}}

\item{RunAssignmentChecks}{\code{logical} Indicating whether assignment checks should be performed or omitted for reduced execution time - Default: \code{TRUE}}

\item{RunSeparately}{\code{logical} - Indicating whether time-consuming functions should be run separately for each server. This can be done for testing purposes or if timeout issues arise. - Default: \code{FALSE}}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}

\item{DS.async}{\code{logical} - Value of argument 'async' in \code{DSI::datashield.assign()} / \code{DSI::datashield.aggregate()} - Default: \code{dsFredaClient::Set.DSSettings$DS.async}}
}
\value{
A \code{list} of report objects
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Perform a complete data preprocessing run of FREDA and CCPhos functionality with default settings.
}
\author{
Bastian Reiter
}
