% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckServerRequirements.R
\name{CheckServerRequirements}
\alias{CheckServerRequirements}
\title{CheckServerRequirements}
\usage{
CheckServerRequirements(
  CCPSiteSpecifications = NULL,
  ServerRequirements = dsCCPhosClient::Meta_ServerRequirements,
  DataSources = NULL
)
}
\arguments{
\item{CCPSiteSpecifications}{\code{data.frame} | Same data frame used for login. Used here only for akquisition of site-specific project names (in case they are differing). | Default: NULL for virtual project}

\item{ServerRequirements}{A list of data frames as defined in \code{\link{Meta_ServerRequirements}}.}

\item{DataSources}{List of DSConnection objects}
}
\value{
A list of messages
}
\description{
Check if technical requirements are met on every participating CCP server.
}
\author{
Bastian Reiter
}
