% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckServerRequirements.R
\name{CheckServerRequirements}
\alias{CheckServerRequirements}
\title{CheckServerRequirements}
\usage{
CheckServerRequirements(
  DataSources = NULL,
  ServerRequirements = dsCCPhosClient::Meta_ServerRequirements
)
}
\arguments{
\item{DataSources}{List of DSConnection objects}

\item{ServerRequirements}{A list of data frames as defined in \code{\link{Meta_ServerRequirements}}.}
}
\value{
A list of messages
}
\description{
Check if technical requirements are met on every participating CCP server.
}
\author{
Bastian Reiter
}
