% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToVirtualCCP.R
\name{ConnectToVirtualCCP}
\alias{ConnectToVirtualCCP}
\title{ConnectToVirtualCCP}
\usage{
ConnectToVirtualCCP(
  CCPTestData,
  NumberOfSites = 1,
  NumberOfPatientsPerSite = NULL
)
}
\arguments{
\item{CCPTestData}{List of CCP test data}

\item{NumberOfSites}{Integer | The number of virtual sites to install}

\item{NumberOfPatientsPerSite}{Integer | Optional value to restrict size of data set for faster testing | Default: NULL}
}
\value{
A list of DSConnection-objects
}
\description{
Sets up a virtual dataSHIELD infrastructure that enables trying out real dsCCPhos functionality on test data.
}
\author{
Bastian Reiter
}
