% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.AugmentData.R
\name{ds.AugmentData}
\alias{ds.AugmentData}
\title{ds.AugmentData}
\usage{
ds.AugmentData(
  CurationOutputName = "CurationOutput",
  OutputName = "AugmentationOutput",
  DataSources = NULL
)
}
\arguments{
\item{OutputName}{String | Name of output object to be assigned on server | Default: 'AugmentationOutput'}

\item{DataSources}{List of DSConnection objects}

\item{CuratedDataSetName}{String | Name of the Curated Data Set object on server | Default: 'CuratedDataSet'}
}
\value{
A list of variables containing messages about object assignment for monitoring purposes.
}
\description{
Transforms curated CCP core data set (CDM) into augmented data set (ADM)
}
\details{
Linked to server-side ASSIGN method AugmentDataDS()
}
\author{
Bastian Reiter
}
