% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.AugmentData.R
\name{ds.AugmentData}
\alias{ds.AugmentData}
\title{ds.AugmentData}
\usage{
ds.AugmentData(
  Name_CurationOutput = "CurationOutput",
  Name_Output = "AugmentationOutput",
  DataSources = NULL
)
}
\arguments{
\item{Name_CurationOutput}{String | Name of curation output object (list) on server | Default: 'CurationOutput'}

\item{Name_Output}{String | Name of assigned output object on server | Default: 'AugmentationOutput'}

\item{DataSources}{List of DSConnection objects}
}
\value{
A list of variables containing messages about object assignment for monitoring purposes.
}
\description{
Transforms curated CCP core data set (CDM) into augmented data set (ADM)
}
\details{
Linked to server-side ASSIGN method AugmentDataDS()
}
\author{
Bastian Reiter
}
