% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.CCP.AugmentData.R
\name{ds.CCP.AugmentData}
\alias{ds.CCP.AugmentData}
\title{ds.CCP.AugmentData}
\usage{
ds.CCP.AugmentData(
  CuratedDataSetName = "CCP.CuratedDataSet",
  OutputName = "CCP.AugmentationOutput",
  RunAssignmentChecks = TRUE,
  UnpackAugmentedDataSet = TRUE,
  RunSeparately = FALSE,
  DSConnections = NULL,
  DS.async = dsFredaClient::Set.DSSettings$DS.async
)
}
\arguments{
\item{CuratedDataSetName}{\code{string} - Name of the Curated Data Set object on server - Default: 'CCP.CuratedDataSet'}

\item{OutputName}{\code{string} - Name of output object to be assigned on server - Default: 'CCP.AugmentationOutput'}

\item{RunAssignmentChecks}{\code{logical} Indicating whether assignment checks should be performed or omitted for reduced execution time - Default: \code{TRUE}}

\item{UnpackAugmentedDataSet}{\code{logical} indicating whether the Augmented Data Set \code{list} should be unpacked so that tables \code{data.frames} are directly accessible - Default: \code{TRUE}}

\item{RunSeparately}{\code{logical} - Indicating whether \code{CurateDataDS()} should be run separately for each server. This can be done for testing purposes or if timeout issues arise. - Default: \code{FALSE}}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}

\item{DS.async}{\code{logical} - Value of argument 'async' in \code{DSI::datashield.assign()} / \code{DSI::datashield.aggregate()} - Default: \code{dsFredaClient::Set.DSSettings$DS.async}}
}
\value{
A \code{list} of variables containing messages about object assignment for monitoring purposes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Transforms curated data set (CDS) into augmented data set (ADS)
}
\details{
Linked to server-side ASSIGN method \code{AugmentDataDS()}
}
\author{
Bastian Reiter
}
