% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.CurateData.R
\name{ds.CurateData}
\alias{ds.CurateData}
\title{ds.CurateData}
\usage{
ds.CurateData(
  Name_Output = "CurationOutput",
  Name_RawDataSet = "RawDataSet",
  RuleProfile_DiagnosisAssociation = "Default",
  RuleProfile_DiagnosisRedundancy = "Default",
  DataSources = NULL
)
}
\arguments{
\item{Name_Output}{String | Name of assigned output object on server | Default: 'CurationOutput'}

\item{Name_RawDataSet}{String | Name of Raw Data Set object (list) on server | Default: 'RawDataSet'}

\item{DataSources}{List of DSConnection objects}
}
\value{
A list of variables containing messages about object assignment for monitoring purposes.
}
\description{
Triggers transformation of Raw Data Set (RDS) into Curated Data Set (CDS) on server.
}
\details{
Linked to server-side ASSIGN method CurateDataDS()
}
