% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.CurateData.R
\name{ds.CurateData}
\alias{ds.CurateData}
\title{ds.CurateData}
\usage{
ds.CurateData(
  RawDataSetName = "RawDataSet",
  Settings = list(DataHarmonization = list(Profile = "Default"), FeatureObligations =
    list(Profile = "Default"), FeatureTracking = list(Profile = "Default")),
  OutputName = "CurationOutput",
  DataSources = NULL
)
}
\arguments{
\item{RawDataSetName}{\code{character} - Name of Raw Data Set object (list) on server - Default: 'RawDataSet'}

\item{Settings}{\code{list} - Settings passed to function
\itemize{  \item DataHarmonization \code{list}
\itemize{\item Run \code{logical} - Whether or not to perform data harmonization - Default: \code{TRUE}
\item RuleSet \code{data.frame} - Default: \code{dsCCPhos::Meta_DataHarmonization}
\item Profile \code{character} - Profile name defining rule set to be used for data harmonization. Profile name must be stated in \code{DataHarmonization$RuleSet} - Default: 'Default'}
\item FeatureObligations \code{list}
\itemize{\item RuleSet \code{data.frame} - Default: \code{dsCCPhos::Meta_FeatureObligations}
\item Profile \code{character} - Profile name defining strict and trans-feature rules for obligatory feature content. Profile name must be stated in \code{FeatureObligations$RuleSet} - Default: 'Default'}
\item FeatureTracking \code{list}
\itemize{\item RuleSet \code{data.frame} - Default: \code{dsCCPhos::Meta_FeatureTracking}
\item Profile \code{character} - Profile name defining which features should be tracked/monitored during curation process. Profile name must be stated in \code{FeatureTracking$RuleSet} - Default: 'Default'}
\item TableCleaning \code{list}
\itemize{\item Run \code{logical} - Whether or not to perform table cleaning (removal of redundant and ineligible entries) - Default: \code{TRUE}}}}

\item{OutputName}{\code{character} - Name of output object to be assigned on server - Default: 'CurationOutput'}

\item{DataSources}{\code{list} of DSConnection objects}
}
\value{
\code{list} of following objects:
\itemize{\item 'Messages' - Info messages concerning completion of \code{CurateDataDS()} and assignment of the following objects on server:
\itemize{\item CurationOutput (\code{list})
\itemize{ \item CuratedDataSet \code{list}
\itemize{ \item BioSampling
\item Diagnosis
\item DiseaseStatus
\item GeneralCondition
\item Histology
\item Metastasis
\item MolecularDiagnostics
\item OtherClassification
\item Patient
\item RadiationTherapy
\item Staging
\item Surgery
\item SystemicTherapy
\item TherapyRecommendation}
\item CurationReport \code{list}
\itemize{\item EntryCounts \code{tibble}
\itemize{ \item Table
\item InitialCount
\item ExcludedPrimary
\item AfterPrimaryExclusion
\item ExcludedSecondary
\item AfterSecondaryExclusion
\item ExcludedSecondaryRedundancy
\item AfterSecondaryRedundancyExclusion}
\item Transformation (list of lists)
\itemize{ \item Monitors
\item EligibilityOverviews
\item ValueSetOverviews}}
\item CurationMessages \code{list}}}
\item 'CurationCompletionCheck'}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Triggers transformation of Raw Data Set (RDS) into Curated Data Set (CDS) on server.
}
\details{
Linked to server-side ASSIGN methods dsCCPhos::CurateDataDS() and dsCCPhos::ExtractFromListDS()
}
\author{
Bastian Reiter
}
