% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.CurateData.R
\name{ds.CurateData}
\alias{ds.CurateData}
\title{ds.CurateData}
\usage{
ds.CurateData(
  RawDataSetName = "RawDataSet",
  OutputName = "CurationOutput",
  RuleProfile_RawDataTransformation = "Default",
  RuleProfile_DiagnosisRedundancy = "Default",
  RuleProfile_DiagnosisAssociation = "Default",
  DataSources = NULL
)
}
\arguments{
\item{RawDataSetName}{String | Name of Raw Data Set object (list) on server | Default: 'RawDataSet'}

\item{OutputName}{String | Name of output object to be assigned on server | Default: 'CurationOutput'}

\item{DataSources}{List of DSConnection objects}
}
\value{
Info messages concerning completion of CurateDataDS() and assignment of the following objects on server:
\itemize{\item CurationOutput (list)
\item CuratedDataSet (list)
\item CurationReport (list)
\item CurationMessages (list)}
}
\description{
Triggers transformation of Raw Data Set (RDS) into Curated Data Set (CDS) on server.
}
\details{
Linked to server-side ASSIGN methods dsCCPhos::CurateDataDS() and dsCCPhos::ExtractFromListDS()
}
\author{
Bastian Reiter
}
