% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetCohortDescription.R
\name{ds.GetCohortDescription}
\alias{ds.GetCohortDescription}
\title{ds.GetCohortDescription}
\usage{
ds.GetCohortDescription(
  DataSources = NULL,
  DataSetName = "AugmentedDataSet",
  CCPDataSetType = "ADS"
)
}
\arguments{
\item{DataSources}{\code{list} of \code{DSConnection} objects}

\item{DataSetName}{\code{string} - Name of Data Set object (list) on server, usually "RawDataSet", "CuratedDataSet" or "AugmentedDataSet" - Default: "AugmentedDataSet"}

\item{CCPDataSetType}{\code{string} - Indicating the type of CCP data set that should be described, one of "RDS" / "CDS" / "ADS" - Default: "ADS"}
}
\value{
A \code{list} containing descriptive statistics characterizing patient cohort
\itemize{\item }
}
\description{
Obtain site-specific and aggregated descriptive characteristics about the patient cohort to be analyzed
}
\details{
Linked to server-side \code{AGGREGATE} function \code{dsCCPhos::GetCohortDescriptionDS()}.
}
\author{
Bastian Reiter
}
