% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetCrossTable.R
\name{ds.GetCrossTab}
\alias{ds.GetCrossTab}
\title{ds.GetCrossTab}
\usage{
ds.GetCrossTab(TableName, FeatureNames, DSConnections = NULL)
}
\arguments{
\item{TableName}{\code{string} - Name of \code{data.frame} on server}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}

\item{FilterExpression}{\code{string} - \code{dplyr::filter} expression as string}

\item{GroupBy}{\code{string} - Optional \code{dplyr::group_by} expression as string}

\item{OutputName}{\code{string} - Name of resulting \code{data.frame} on server}
}
\value{
A \code{list} of messages about object assignment for monitoring purposes
}
\description{
Get cross tab
}
\details{
Linked to server-side \code{AGGREGATE} function \code{GetCrossTabDS()}
}
\author{
Bastian Reiter
}
