% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetCurationReport.R
\name{ds.GetCurationReport}
\alias{ds.GetCurationReport}
\title{ds.GetCurationReport}
\usage{
ds.GetCurationReport(DataSources = NULL)
}
\arguments{
\item{DataSources}{List of DSConnection objects}
}
\value{
A list of Curation Reports
}
\description{
Receive Curation Reports from sites and create cumulated reporting objects.
}
\details{
Linked to server-side AGGREGATE method GetReportingObjectDS()
}
\author{
Bastian Reiter
}
