% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetCurationReport.R
\name{ds.GetCurationReport}
\alias{ds.GetCurationReport}
\title{ds.GetCurationReport}
\usage{
ds.GetCurationReport(
  Name_CurationOutput = "CurationOutput",
  DataSources = NULL
)
}
\arguments{
\item{Name_CurationOutput}{String | Name of object on server previously assigned by dsCCPhos::CurateData() | Default: 'CurationOutput'}

\item{DataSources}{List of DSConnection objects}
}
\description{
What it does
}
\details{
Linked to server-side AGGREGATE method GetCurationReportDS()
}
