% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetRDSValidationReport.R
\name{ds.GetRDSValidationReport}
\alias{ds.GetRDSValidationReport}
\title{ds.GetRDSValidationReport}
\usage{
ds.GetRDSValidationReport(DataSources = NULL, RawDataSetName = "RawDataSet")
}
\arguments{
\item{DataSources}{List of DSConnection objects}

\item{RawDataSetName}{String | Name of Raw Data Set object (list) on server | Default: 'RawDataSet'}
}
\value{
A list of tibbles containing output of validation
}
\description{
Triggers validation of Raw Data Set on server and requests a report.
}
\details{
Linked to server-side AGGREGATE method GetRDSValidationReportDS()
}
\author{
Bastian Reiter
}
