% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.UnpackAugmentedDataSet.R
\name{ds.UnpackAugmentedDataSet}
\alias{ds.UnpackAugmentedDataSet}
\title{ds.UnpackAugmentedDataSet}
\usage{
ds.UnpackAugmentedDataSet(
  AugmentedDataSetName = "AugmentedDataSet",
  DataSources = NULL
)
}
\arguments{
\item{AugmentedDataSetName}{String | Name of Augmented Data Set object (list) on server | Default: 'AugmentedDataSet'}

\item{DataSources}{List of DSConnection objects}
}
\value{
A list of messages about object assignment for monitoring purposes
}
\description{
Make tables within Augmented Data Set (list object) directly addressable in R server sessions
}
\details{
Linked to server-side ASSIGN method ExtractFromListDS()
}
\author{
Bastian Reiter
}
