% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.UnpackCuratedDataSet.R
\name{ds.UnpackCuratedDataSet}
\alias{ds.UnpackCuratedDataSet}
\title{ds.UnpackCuratedDataSet}
\usage{
ds.UnpackCuratedDataSet(
  CuratedDataSetName = "CuratedDataSet",
  DataSources = NULL
)
}
\arguments{
\item{CuratedDataSetName}{\code{string} - Name of Curated Data Set object (list) on server | Default: 'CuratedDataSet'}

\item{DataSources}{List of DSConnection objects}
}
\value{
A list of messages about object assignment for monitoring purposes
}
\description{
Make tables within Curated Data Set (list object) directly addressable in R server sessions
}
\details{
Linked to server-side ASSIGN method ExtractFromListDS()
}
\author{
Bastian Reiter
}
