% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetServerWorkspaceInfo.R
\name{GetServerWorkspaceInfo}
\alias{GetServerWorkspaceInfo}
\title{GetServerWorkspaceInfo}
\usage{
GetServerWorkspaceInfo(DataSources = NULL)
}
\arguments{
\item{DataSources}{\code{list} of \code{DSConnection} objects}
}
\value{
A list
}
\description{
Check which objects live in server-side R sessions and collect meta data about them.
}
\author{
Bastian Reiter
}
