% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadRawDataSet.R
\name{LoadRawDataSet}
\alias{LoadRawDataSet}
\title{LoadRawDataSet}
\usage{
LoadRawDataSet(SiteSpecifications = NULL, DSConnections = NULL)
}
\arguments{
\item{SiteSpecifications}{\code{data.frame} - Same \code{data.frame} used for login. Used here only for acquisition of site-specific project names (in case they are differing) - Default: \code{NULL} for virtual project}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} of messages
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Load raw data set from Opal data base into R session on servers.
}
\author{
Bastian Reiter
}
