% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadRawDataSet.R
\name{LoadRawDataSet}
\alias{LoadRawDataSet}
\title{LoadRawDataSet}
\usage{
LoadRawDataSet(CCPSiteSpecifications = NULL, DataSources)
}
\arguments{
\item{CCPSiteSpecifications}{\code{data.frame} | Same data frame used for login. Used here only for akquisition of site-specific project names (in case they are differing). | Default: NULL for virtual project}

\item{DataSources}{\code{list} of \code{DSConnection} objects}
}
\value{
A \code{list} of messages
}
\description{
Load raw data set from Opal data base into R session on servers.
}
\author{
Bastian Reiter
}
